Fix barters loading as nothing, and stop a dead source failing silently
Barters shipped empty in 0.1.0. tarkov.dev records currency-priced trades
with fractional counts — a barter costing "155.1" of something is real data,
and 313 of the 789 carry one — so reading counts as integers threw on the
whole document and every item came back with no trades against it.

The reason it shipped that way is the more interesting bug. A fetch that
failed was turned into null so one dead source could not take the rest down
with it, which is right, but nothing recorded that it had failed. The
refresh reported success, status looked healthy, and a whole data source was
missing with no way to tell. Failing soft without saying so is just failing.

Each source is now named as it is fetched, failures are recorded against
that name and cleared when it recovers, and Setup goes red naming what is
down rather than showing a comfortable green tick over missing data.

160 tests, including the two that would have caught this.

+++ src/RatNav.Core/Data/TarkovDevClient.cs
@@ -369,7 +369,10 @@ public sealed class TarkovDevClient(HttpClient http)
         List<BarterItemDto>? RequiredItems,
         BarterItemDto? OfferedItem);
 
-    private sealed record BarterItemDto(string? Item, int Count);
+    // Counts are fractional. A barter that costs "155.1" of something is not a rounding error in
+    // the source — it is how tarkov.dev records trades priced in currency, and 313 of the 789
+    // barters have one. Reading them as integers failed the whole document.
+    private sealed record BarterItemDto(string? Item, double Count);
 
     private sealed record MapDto(
         string? Name, string? NormalizedName, string? NameId, List<ExtractDto>? Extracts);
